requirements.txt
Flask
google-cloud-aiplatform
google-cloud-storage
python-dotenv
